refactor: Money/Currency 値オブジェクトを estimate から shared/domain/values へ昇格
Money は estimate 固有でなく金額を扱う任意の文脈で再利用しうる汎用 VO のため、
既存の汎用 VO（Address/CompanyCode 等）と同じ shared 層へ移す。Currency は Money の
実依存先で外部直接利用者が無いため、依存逆転（shared→subdomain）を避けるべく同時移動する。
挙動・公開シグネチャ・JSDoc は不変（4ファイルは R100 の純粋リネーム、差分は import パスのみ）。

移動と import 更新を分けると中間状態が tsc を通らず、Money/Currency も同時移動が必須で
緑を保てる中間点が無いため、移設＋被参照33件の追従を単一の原子的コミットにまとめた。

#426 pricing の SellingUnitPrice が shared Money に依存するための先行前提。
JSDoc の estimate スペック参照整理（§8.1・掛率・税率）は #438 へ切り出し、本 PR の純度を保つ。
