+++ prisma/seed-estimates.ts
@@ -29,7 +29,7 @@ import { DiscountRate } from "@subdomains/estimate/domain/values/DiscountRate";
 import { EstimateNumber } from "@subdomains/estimate/domain/values/EstimateNumber";
 import { FaultDescription } from "@subdomains/estimate/domain/values/FaultDescription";
 import { ItemName } from "@subdomains/estimate/domain/values/ItemName";
-import { Money } from "@subdomains/estimate/domain/values/Money";
+import { Money } from "@server/shared/domain/values/Money";
 import { Quantity } from "@subdomains/estimate/domain/values/Quantity";
 import { SubmissionType } from "@subdomains/estimate/domain/values/SubmissionType";
 import { TaxRate } from "@subdomains/estimate/domain/values/TaxRate";
