+++ src/server/subdomains/estimate/domain/entities/__tests__/RevisedEstimateItemDetail.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { Money } from "../../values/Money";
+import { Money } from "@server/shared/domain/values/Money";
 import { RevisedEstimateItemDetailId } from "../../values/RevisedEstimateItemDetailId";
 import { RevisedEstimateItemDetail } from "../RevisedEstimateItemDetail";
