+++ src/server/subdomains/estimate/domain/policies/SetGroupDerivationPolicy.ts
@@ -1,5 +1,5 @@
 import { BusinessRuleViolationError } from "@server/shared/errors/DomainError";
-import { Money } from "../values/Money";
+import { Money } from "@server/shared/domain/values/Money";
 
 /**
  * 導出に必要な構成明細の射影。EstimateItem はこの構造に適合する（finalAmount / sortOrder
