+++ src/server/subdomains/estimate/domain/services/EstimateDuplicationService.ts
@@ -10,7 +10,7 @@ import {
 } from "../entities";
 import { EstimateNumber } from "../values/EstimateNumber";
 import { EstimateVariationId } from "../values/EstimateVariationId";
-import { Money } from "../values/Money";
+import { Money } from "@server/shared/domain/values/Money";
 import { TaxRate } from "../values/TaxRate";
 
 /**
