+++ src/server/subdomains/estimate/domain/services/__tests__/EstimateDuplicationService.test.ts
@@ -13,7 +13,7 @@ import { EstimateVariationId } from "../../values/EstimateVariationId";
 import { FaultDescription } from "../../values/FaultDescription";
 import { ItemName } from "../../values/ItemName";
 import { Memo } from "../../values/Memo";
-import { Money } from "../../values/Money";
+import { Money } from "@server/shared/domain/values/Money";
 import { Quantity } from "../../values/Quantity";
 import { SubmissionType } from "../../values/SubmissionType";
 import { TaxRate } from "../../values/TaxRate";
